refactor/javascript functions into seperate service files
